[refs #00267] Migrate sky.com/toolkit documentation to docs/
Currently, Toolkit's documentation is hosted on an external source which
has proved difficult to maintain. Ideally, we want to keep our
documentation as close to the code as possible.

This commit migrates the **existing** documentation for:

* buttons
* calendar
* divider
* dropdown
* forms
* hero
* links
* panel
* select
* shine
* spinner
* tables
* tile
* tooltip

Example images added to sky.com/assets/toolkit/docs for each component,
along with JSON at the end of each code block to render specific
example templates.